requirements.txt
beautifulsoup4==4.6.0
Django==1.11
django-crontab==0.7.1
django-debug-toolbar==1.9.1
django-mptt==0.9.0
ipython==6.4.0
mysqlclient==1.3.13
pkg-resources==0.0.0
pytz==2018.5
requests==2.19.1
